refactor: dirty check 역할을 EntityPersister으로 이관
