+++ src/main/java/persistence/entity/DefaultEntityPersister.java
@@ -5,8 +5,17 @@ import jdbc.JdbcTemplate;
 import persistence.sql.dml.DeleteQueryBuilder;
 import persistence.sql.dml.InsertQueryBuilder;
 import persistence.sql.dml.UpdateQueryBuilder;
+import persistence.sql.meta.EntityColumn;
+import persistence.sql.meta.EntityTable;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class DefaultEntityPersister implements EntityPersister {
+    public static final String NOT_CHANGED_MESSAGE = "변경된 필드가 없습니다.";
+    
     private final JdbcTemplate jdbcTemplate;
     private final InsertQueryBuilder insertQueryBuilder;
     private final UpdateQueryBuilder updateQueryBuilder;
@@ -28,11 +37,30 @@ public class DefaultEntityPersister implements EntityPersister {
 
     @Override
     public void update(Object entity, Object snapshot) {
-        jdbcTemplate.execute(updateQueryBuilder.update(entity, snapshot));
+        final EntityTable entityTable = new EntityTable(entity);
+        final EntityTable snapshotEntityTable = new EntityTable(snapshot);
+        final List<EntityColumn> dirtiedEntityColumns = getDirtiedEntityColumns(entityTable, snapshotEntityTable);
+        jdbcTemplate.execute(updateQueryBuilder.update(entityTable, dirtiedEntityColumns));
     }
 
     @Override
     public void delete(Object entity) {
         jdbcTemplate.execute(deleteQueryBuilder.delete(entity));
+    }
+
+    private List<EntityColumn> getDirtiedEntityColumns(EntityTable entityTable, EntityTable snapshotEntityTable) {
+        final List<EntityColumn> dirtiedEntityColumns = IntStream.range(0, entityTable.getColumnCount())
+                .filter(i -> isDirtied(entityTable.getEntityColumn(i), snapshotEntityTable.getEntityColumn(i)))
+                .mapToObj(entityTable::getEntityColumn)
+                .collect(Collectors.toList());
+
+        if (dirtiedEntityColumns.isEmpty()) {
+            throw new IllegalStateException(NOT_CHANGED_MESSAGE);
+        }
+        return dirtiedEntityColumns;
+    }
+
+    private boolean isDirtied(EntityColumn entityColumn, EntityColumn snapshotEntityColumn) {
+        return !Objects.equals(entityColumn.getValue(), snapshotEntityColumn.getValue());
     }
 }
